// frontend/src/api/models/customerResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * backend-node
 * OpenAPI spec version: 1.0.0
 */

export interface CustomerResponse {
  id: number;
  nome: string;
  cognome: string;
  email: string;
  indirizzo?: string;
  citta?: string;
  cap?: string;
  telefono: string;
  telefonoSecondario?: string;
}